// angularProject/src/app/api/model/productDto.ts
/**
 * TrueValue
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface ProductDto { 
    id?: number;
    name?: string | null;
    shortDescription?: string | null;
    fullDescription?: string | null;
    productTypeId?: number;
    parentGroupedProductId?: number;
    visibleIndividually?: boolean;
    adminComment?: string | null;
    productTemplateId?: number;
    vendorId?: number;
    showOnHomepage?: boolean;
    metaKeywords?: string | null;
    metaDescription?: string | null;
    metaTitle?: string | null;
    allowCustomerReviews?: boolean;
    approvedRatingSum?: number;
    notApprovedRatingSum?: number;
    approvedTotalReviews?: number;
    notApprovedTotalReviews?: number;
    subjectToAcl?: boolean;
    limitedToStores?: boolean;
    sku?: string | null;
    manufacturerPartNumber?: string | null;
    gtin?: string | null;
    isGiftCard?: boolean;
    giftCardTypeId?: number;
    overriddenGiftCardAmount?: number | null;
    requireOtherProducts?: boolean;
    requiredProductIds?: string | null;
    automaticallyAddRequiredProducts?: boolean;
    isDownload?: boolean;
    downloadId?: number;
    unlimitedDownloads?: boolean;
    maxNumberOfDownloads?: number;
    downloadExpirationDays?: number | null;
    downloadActivationTypeId?: number;
    hasSampleDownload?: boolean;
    sampleDownloadId?: number;
    hasUserAgreement?: boolean;
    userAgreementText?: string | null;
    isRecurring?: boolean;
    recurringCycleLength?: number;
    recurringCyclePeriodId?: number;
    recurringTotalCycles?: number;
    isRental?: boolean;
    rentalPriceLength?: number;
    rentalPricePeriodId?: number;
    isShipEnabled?: boolean;
    isFreeShipping?: boolean;
    shipSeparately?: boolean;
    additionalShippingCharge?: number;
    deliveryDateId?: number;
    isTaxExempt?: boolean;
    taxCategoryId?: number;
    isTelecommunicationsOrBroadcastingOrElectronicServices?: boolean;
    manageInventoryMethodId?: number;
    productAvailabilityRangeId?: number;
    useMultipleWarehouses?: boolean;
    warehouseId?: number;
    stockQuantity?: number;
    displayStockAvailability?: boolean;
    displayStockQuantity?: boolean;
    minStockQuantity?: number;
    lowStockActivityId?: number;
    notifyAdminForQuantityBelow?: number;
    backorderModeId?: number;
    allowBackInStockSubscriptions?: boolean;
    orderMinimumQuantity?: number;
    orderMaximumQuantity?: number;
    allowedQuantities?: string | null;
    allowAddingOnlyExistingAttributeCombinations?: boolean;
    notReturnable?: boolean;
    disableBuyButton?: boolean;
    disableWishlistButton?: boolean;
    availableForPreOrder?: boolean;
    preOrderAvailabilityStartDateTimeUtc?: string | null;
    callForPrice?: boolean;
    price?: number;
    oldPrice?: number;
    productCost?: number;
    customerEntersPrice?: boolean;
    minimumCustomerEnteredPrice?: number;
    maximumCustomerEnteredPrice?: number;
    basepriceEnabled?: boolean;
    basepriceAmount?: number;
    basepriceUnitId?: number;
    basepriceBaseAmount?: number;
    basepriceBaseUnitId?: number;
    markAsNew?: boolean;
    markAsNewStartDateTimeUtc?: string | null;
    markAsNewEndDateTimeUtc?: string | null;
    hasTierPrices?: boolean;
    hasDiscountsApplied?: boolean;
    weight?: number;
    length?: number;
    width?: number;
    height?: number;
    availableStartDateTimeUtc?: string | null;
    availableEndDateTimeUtc?: string | null;
    displayOrder?: number;
    published?: boolean;
    deleted?: boolean;
    createdOnUtc?: string;
    updatedOnUtc?: string;
    distance?: number | null;
    category?: number;
    categoryType?: string | null;
    actualPrice?: number;
    inItQuantity?: string | null;
    productQuantity?: string | null;
}